refactor(clawdstrike-policy-event): extract util from edr/mod.rs
Move the shared ID/hash/path utility cluster (stable_id, FNV consts,
normalize_path_string/normalize_hostname/hostname_from_url_like,
evidence_hash_for_value, insert_json, telemetry_privacy_report_id_*,
response_execution_id_*/transition_id, reconstruct_path, and the local
ResponseExecutionEffectBindingEntry) into edr/util.rs.

stable_id is the shared seam (deception, simulation, response, ids,
process, receipt/evidence, detection/supply_chain reach it via super::
paths), so every moved item is re-exported pub(crate) at the edr root to
preserve those paths. Private fns widened to pub(crate) only as needed
for the re-export. The 3 flight-recorder schema-version consts stay in
mod.rs (flight_recorder reaches them via super::super::). Pure
relocation; bodies unchanged.
